pyproject.toml
[tool.poetry]
name = "imagetowordlemosaicapi"
version = "0.1.0"
description = ""
authors = ["ThorneWolf <thornecanyon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.2"
fastapi = "^0.74.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
